Avoid freeing block_handler too early
Client::new would automatically free the block_handler, which would also
free the closure. This was initially thought to be fine as es_new_client
should be cloning the block. However, the ConcreteBlock implementation
of copy does not actually do a deep clone - instead it merely does a
memmove of the closure data.

As such, we need to make sure we don't run the destructore on
block_handler.