Greatly simplify ffi_wrap_enum
Pushes the responsibility up to the user to provide the cfg and ident
names, which greatly reduces the amount of repetition in the macro to
handle all the macos versions.

This removes the LAST_<MACOS_VERSION>* variants from the enum. Those
were unused and had little reason to exist (the cfg, along with
auto-docs, provide information on which enums are available on which
versions).

If [macro_metavar_expr_concat] is ever stabilized, we'll be able to
reintroduce it without issues.

[macro_metavar_expr_concat]: https://github.com/rust-lang/rust/issues/124225